# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger-demo",
        "description": "Swagger `Api of ficde`",
        "contact": {
            "email": "swagger@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/ficde"
        },
        {
            "url": "https://ficde.herokuapp.com/ficde"
        }
    ],
    "paths": {
...
        "/asistencias/{dni}": {
            "delete": {
                "tags": [
                    "asistencias"
                ],
                "summary": "Removes a asistencias resource",
                "description": "Removes a asistencias resource",
                "operationId": "App\\Http\\Controllers\\AsistenciasController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The asistencias resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a asistencias resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/cursos/{dni}": {
            "delete": {
                "tags": [
                    "cursos"
                ],
                "summary": "Removes a cursos resource",
                "description": "Removes a cursos resource",
                "operationId": "App\\Http\\Controllers\\CursosController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The cursos resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a cursos resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/estudiantes/{dni}": {
            "get": {
                "tags": [
                    "estudiantes"
                ],
                "summary": "Gets a estudiante resource",
                "description": "Gets a estudiante resource",
                "operationId": "App\\Http\\Controllers\\EstudiantesController::show",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The estudiante resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a estudiante resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "estudiantes"
                ],
                "summary": "Updates a estudiantes resource",
                "description": "Updates a estudiantes resource by the dni",
                "operationId": "App\\Http\\Controllers\\EstudiantesController::update",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "estudiantes resource id",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a estudiantes resource",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/estudiantes"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The estudiantes resource updated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "estudiantes"
                ],
                "summary": "Removes a estudiantes resource",
                "description": "Removes a estudiantes resource",
                "operationId": "App\\Http\\Controllers\\EstudiantesController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The estudiantes resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a estudiantes resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/estudiantes_materias/{dni}": {
            "delete": {
                "tags": [
                    "estudiantes_materias"
                ],
                "summary": "Removes a estudiantes_materias resource",
                "description": "Removes a estudiantes_materias resource",
                "operationId": "App\\Http\\Controllers\\Estudiantes_MateriasController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The estudiantes_materias resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a estudiantes_materias resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/institucion/{dni}": {
            "delete": {
                "tags": [
                    "institucion"
                ],
                "summary": "Removes a institucion resource",
                "description": "Removes a institucion resource",
                "operationId": "App\\Http\\Controllers\\InstitucionController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The institucion resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a institucion resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/justificados/{dni}": {
            "delete": {
                "tags": [
                    "justificados"
                ],
                "summary": "Removes a justificados resource",
                "description": "Removes a justificados resource",
                "operationId": "App\\Http\\Controllers\\JustificadosController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The justificados resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a justificados resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/materias/{dni}": {
            "delete": {
                "tags": [
                    "materias"
                ],
                "summary": "Removes a materias resource",
                "description": "Removes a materias resource",
                "operationId": "App\\Http\\Controllers\\MateriasController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The materias resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a materias resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/modalidades/{dni}": {
            "delete": {
                "tags": [
                    "modalidades"
                ],
                "summary": "Removes a modalidades resource",
                "description": "Removes a modalidades resource",
                "operationId": "App\\Http\\Controllers\\ModalidadesController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The modalidades resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a modalidades resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
        "/notas/{dni}": {
            "delete": {
                "tags": [
                    "notas"
                ],
                "summary": "Removes a notas resource",
                "description": "Removes a notas resource",
                "operationId": "App\\Http\\Controllers\\NotasController::destroy",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "The notas resource dni",
                        "required": true,
                        "schema": {
                            "description": "The unique identifier of a notas resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "The resource has been deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "estudiantes": {
                "properties": {
                    "nombre": {
                        "description": "The Estudiantes nombre",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "apellido": {
                        "description": "The Estudiantes apellido",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "dni": {
                        "description": "The Estudiantes dni",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "email": {
                        "description": "The Estudiantes email",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "telefono": {
                        "description": "The Estudiantes telefono",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "direccion": {
                        "description": "The Estudiantes direccion",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "status": {
                        "description": "The asistencias telefono",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "user": {
                        "description": "The aulas user",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}